add Cart

+++ Zal_User/Controller/Tabbar/Home/NearStore/NearStoreModel.swift
@@ -24,7 +24,7 @@ struct nearModel: Codable {
     let phone, address, lat, lng: String?
     let city, area: Area?
     let mainCategory: MainCategory?
-    let rate: Int?
+    let rate: Double?
     let isVerified: Bool?
     let distance: Double?
